Configure projects to use maven
Visual COBOL 6.0 enables building COBOL JVM projects with maven. The examples projects in this repository were originally configured with Visual COBOL 5.0 where maven support is missing. Hence why, the projects used in the Visual COBOL book "New Application Modernization Tools for the Java Developer" were configured to be built with combination of ant and maven.

With this change all projects are now built using maven only. This includes running COBOL JVM unit tests as part of maven build lifecycle.
The examples code generally remain untouched - some fixes has been made so that the projects are built and run without any errors. The samples can be used by following the Visual COBOL book "New Application Modernization Tools for the Java Developer" considering that ant is no longer used for building. All other configuration steps and guidelines in the book are still applicable.

The example projects have been tested and verified with Visual COBOL 7.0 on Windows and Linux platforms.
Small adjustment should be made when the projects will be used with Visual COBOL 6.0. The java projects that consume COBOL JVM code require RTS jars files to be installed into maven's local repository. Instructions how to do that are added into the pom.xml files of those java projects. The instructions are based on 7.0 version, so when the example projects will be used with Visual COBOL 6.0, the version of RTS jars should be adjusted appropriately.